round of fixes for client concentration
3. Need to delete messages from the client queue once added to queue for
sending
4. On start up the registry shows Error still and status of off when
really It is on and heart beat is updating
5. Restart the connection of the client periodically
7. Figure out why the connection is flapping – test first and send back
the connection socket-(test for connection or server listening and then
try to connect if not connected – write back status of port and if can’t
write then dispose the connection and reconnect and then send messages.
Put in timer and don’t just endlessly connect.
8. Need to update the last message date time in registry
9. Fix queue not updating the garbage file
10. Fix mhistory not updating correctly as well.
11. Fix the multiple open error files error in the code
13. Turn into a service